Move other notebook route
